requirements.txt
pyjwt
passlib[bcrypt]
fastapi
uvicorn
pydantic
psycopg[binary]
python-dotenv
mercadopago